@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <RootNamespace>Leayal.Shared.Windows</RootNamespace>
     <UseWindowsForms>true</UseWindowsForms>
     <Authors>Dramiel Leayal</Authors>
