@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Configurations>Debug;Release;Debug HTTP Requests</Configurations>
     <Authors>Dramiel Leayal</Authors>
     <Nullable>enable</Nullable>
