@@ -8,6 +8,8 @@ namespace Leayal.PSO2Launcher.Core.Classes.PSO2.DataTypes
     // Another madness's gonna be here.
     public class PatchListItem : IEquatable<PatchListItem>
     {
+        internal static readonly PatchListItem Empty = new PatchListItem(null, ReadOnlyMemory<char>.Empty, ReadOnlyMemory<char>.Empty, 0, null);
+
         internal const string AffixFilename = ".pat";
 
         internal readonly PatchListBase? Origin;
@@ -77,7 +79,7 @@ namespace Leayal.PSO2Launcher.Core.Classes.PSO2.DataTypes
             }
         }
 
-        public static ReadOnlySpan<char> GetFilenameWithoutAffix(in ReadOnlySpan<char> filename)
+        public static ReadOnlySpan<char> GetFilenameWithoutAffix(scoped in ReadOnlySpan<char> filename)
         {
             if (filename.EndsWith(AffixFilename, StringComparison.OrdinalIgnoreCase))
             {
