@@ -1,38 +1,41 @@
 ﻿using System;
+using System.Collections.Frozen;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
+using System.Windows.Input;
 using Leayal.PSO2Launcher.Core.Classes.PSO2.DataTypes;
 
 namespace Leayal.PSO2Launcher.Core.Classes.PSO2
 {
     public class PatchListMemory : PatchListBase, IReadOnlyCollection<PatchListItem>
     {
-        private readonly Dictionary<string, PatchListItem> _items;
+        private readonly FrozenDictionary<string, PatchListItem> _items;
 
         public override bool CanCount => true;
 
         public override int Count => this._items.Count;
 
-        public PatchListMemory(PatchRootInfo rootInfo, bool? isReboot, IDictionary<string, PatchListItem> items) : base(rootInfo, isReboot)
+        public PatchListMemory(PatchRootInfo rootInfo, bool? isReboot, IReadOnlyDictionary<string, PatchListItem> items) : base(rootInfo, isReboot)
         {
-            this._items = new Dictionary<string, PatchListItem>(items, PathStringComparer.Default);
+            this._items = FrozenDictionary.ToFrozenDictionary(items, PathStringComparer.Default);
         }
 
         public PatchListMemory(PatchRootInfo rootInfo, bool? isReboot) : base(rootInfo, isReboot)
         {
-            this._items = new Dictionary<string, PatchListItem>(PathStringComparer.Default);
+            this._items = FrozenDictionary<string, PatchListItem>.Empty;
         }
 
-        internal PatchListMemory(PatchRootInfo rootInfo, bool? isReboot, Dictionary<string, PatchListItem> items) : base(rootInfo, isReboot)
+        protected override IEnumerator<PatchListItem> CreateEnumerator()
         {
-            this._items = items;
+            var items = this._items.Values;
+            var count = items.Length;
+            for (int i = 0; i < count; i++) yield return items.ItemRef(i);
         }
-
-        protected override IEnumerator<PatchListItem> CreateEnumerator() => this._items.Values.GetEnumerator();
 
         public override bool TryGetByFilenameExact(string filename, [NotNullWhen(true)] out PatchListItem? value) => this._items.TryGetValue(filename, out value);
 
-        protected override void CopyTo(Dictionary<string, PatchListItem> items, bool clearBeforeCopy)
+        protected override void CopyTo(IDictionary<string, PatchListItem> items, bool clearBeforeCopy)
         {
             if (clearBeforeCopy)
             {
@@ -48,8 +51,14 @@ namespace Leayal.PSO2Launcher.Core.Classes.PSO2
         {
             if (disposing)
             {
-                this._items.Clear();
-                this._items.TrimExcess(0);
+                if (this._items is IDictionary<string, PatchListItem> dict)
+                {
+                    dict.Clear();
+                    if (dict is Dictionary<string, PatchListItem> dict2)
+                    {
+                        dict2.TrimExcess(0);
+                    }
+                }
             }
         }
     }
